src/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
filelock==3.0.12
fsspec==2021.5.0
future==0.18.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
huggingface-hub==0.0.8
idna==2.10
joblib==1.0.1
Markdown==3.3.4
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tokenizers==0.10.2
torchaudio==0.8.0
torchmetrics==0.3.2
tqdm==4.60.0
transformers==4.6.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
yarl==1.6.3
